feat: implement esm.mjs module
