+++ src/modules/esm.mjs
@@ -1,0 +1,47 @@
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const aJsonPath = path.join(__dirname, 'files', 'a.json');
+const bJsonPath = path.join(__dirname, 'files', 'b.json');
+const aJson = await fs.promises.readFile(aJsonPath, 'utf-8');
+const bJson = await fs.promises.readFile(bJsonPath, 'utf-8');
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = aJson;
+} else {
+    unknownObject = bJson;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__dirname}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {
+    unknownObject,
+    myServer,
+};
+
